FoodCard, menu and the UI structure
