+++ ui/src/app/restaurants/[restaurantId]/branches/[branchId]/admin/page.jsx
@@ -1,0 +1,12 @@
+
+
+export default function AdminBranch() {
+    return (
+      <div>
+        <h2>
+          Admin page for a single branch
+        </h2>
+  
+      </div>
+    )
+  }
