+++ ui/src/app/restaurants/[restaurantId]/branches/[branchId]/menu/[foodId]/edit.jsx
@@ -1,0 +1,12 @@
+
+
+export default function Restaurant() {
+    return (
+      <div>
+        <h2>
+          Menu Page can be edited here by admin
+        </h2>
+  
+      </div>
+    )
+  }
