+++ ui/src/app/restaurants/[restaurantId]/branches/[branchId]/menu/[foodId]/page.jsx
@@ -1,0 +1,12 @@
+
+
+export default function Food() {
+    return (
+      <div>
+        <h2>
+          Food Item
+        </h2>
+  
+      </div>
+    )
+  }
