+++ ui/src/app/restaurants/[restaurantId]/branches/page.jsx
@@ -1,0 +1,13 @@
+import Components from "@/components/Components";
+
+export default async function Branches({ params }) {
+    const { restaurantId } = await params;
+  return (
+    <div>
+      <h2>
+        ALL branches page for any restaurant {restaurantId}
+      </h2>
+
+    </div>
+  )
+}
